Restructured CustomListItem
